Added the word adjacent to make it clear we are not talking about the new adjacent ranges, so not the new ranges created by spliting.
changed "artifically inflated" to "set"

and reword the last sentence to use active voice rather then passive. This parahraph is passive heavy which is fine because its still clear and well written, but I just dont want to go an entire parahraph in passive voice.

Also added a todo about the buckets. I wanted to reword that second sentence there just to make it a little clearer but I wanted to make sure I understood it correctly before doing so.