So, I didn't really understand the pertubation section since crossover, mutation, range adapation, and local opts pertube, or try to pertube, the system in some way.
I think maybe the aim was to note that test suite crossover is supposed to be mixing up ranges, while things like test case crossover does not. But it would probally just be easier to mention this where crossover or ranges are discussed.

So, I removed the pertubation section tag but did it on a seperate branch in case i am missing somthing.

I also recombined the two types of crossover into a single section. There is so much overlap between these two operators it seems wasteful to repeat the same stuff in two different sections. I did try to reword it from what it said when they were originally one section, and I think its pretty clear now.