Extended problem description and some other small tweaks
Added explicit formal definitions of CFGs, branch and multiple condition
coverage to the problem description to replace the old generic
description used in the proposal.

Also made a number of small tweaks to try to get things fitting on pages
nicely. We are now at exactly 16 pages of content, with pretty much no
room to spare. THen a 17th page with the appendix and references.

Honestly theres still plenty of problems with this paper that could be
fixed, tweaked, pruned, and revised but I think were long past the point
of diminishing return on this and I'm find with just turning it in as
is.

I was going to fix that citation but I can't even find the paper that
its referring to. Theres one in the sources folder with an extremely
similar name but only has 2 authors. So maybe you just have that one
printed out?